+++ src/resources/index.ts
@@ -24,6 +24,8 @@ export {
   type Message,
   type TogglePublicRequest,
   type TogglePublicResponse,
+  type UpdateChatSessionFavoriteRequest,
+  type UpdateChatSessionRequest,
   type ChatAddGitCommitResponse,
   type ChatCopyNodeOutputByCodeHashResponse,
   type ChatDeleteFilesResponse,
@@ -45,6 +47,7 @@ export {
   type ChatRevertToCommitParams,
   type ChatTogglePublicParams,
   type ChatUpdateSessionParams,
+  type ChatUpdateSessionFavoriteParams,
 } from './chat';
 export {
   Code,
@@ -180,7 +183,13 @@ export {
   type JobListParams,
   type JobStatusParams,
 } from './jobs';
-export { Projects, type DeleteProjectResponse, type GetProjectResponse, type Project } from './projects';
+export {
+  Projects,
+  type DeleteProjectResponse,
+  type Project,
+  type UpdateProjectRequest,
+  type ProjectUpdateParams,
+} from './projects';
 export { PublicSessions } from './public-sessions';
 export {
   Report,
@@ -303,7 +312,6 @@ export {
   type AddMemberRequest,
   type AddMemberResponse,
   type CreateProjectRequest,
-  type CreateProjectResponse,
   type CreateTeamRequest,
   type CreateTeamResponse,
   type CreditsUsageRequest,
@@ -312,6 +320,7 @@ export {
   type DeleteTeamResponse,
   type GetTeamResponse,
   type Granularity,
+  type InvitationDetailsResponse,
   type ListMembersResponse,
   type ListProjectsResponse,
   type ListTeamsResponse,
@@ -320,6 +329,8 @@ export {
   type TeamRole,
   type TeamSubscriptionStatus,
   type TeamWithRole,
+  type UpdateMemberRoleRequest,
+  type UpdateMemberRoleResponse,
   type UpdateTeamRequest,
   type UpdateTeamResponse,
   type UsageGroupKey,
@@ -330,6 +341,7 @@ export {
   type TeamAddMemberParams,
   type TeamCreateProjectParams,
   type TeamCreditsUsageParams,
+  type TeamUpdateMemberRoleParams,
 } from './teams';
 export {
   User,